// src/types/district.ts
export enum Districts {
    THIRUVANANTHAPURAM = "THIRUVANANTHAPURAM",
    KOLLAM = "KOLLAM",
    PATHANAMTHITTA = "PATHANAMTHITTA",
    ALAPPUZHA = "ALAPPUZHA",
    KOTTAYAM = "KOTTAYAM",
    IDUKKI = "IDUKKI",
    ERNAKULAM = "ERNAKULAM",
    THRISSUR = "THRISSUR",
    PALAKKAD = "PALAKKAD",
    MALAPPURAM = "MALAPPURAM",
    KOZHIKODE = "KOZHIKODE",
    WAYANAD = "WAYANAD",
    KANNUR = "KANNUR",
    KASARAGOD = "KASARAGOD",
    OTHERS_IN_INDIA = "OTHERS_IN_INDIA",
    OUTSIDE_INDIA = "OUTSIDE_INDIA",
}